refactor(input): narrow help overlay state boundary
