+++ src/backend/wayland/state/render/ui.rs
@@ -191,7 +191,7 @@ impl WaylandState {
         height: u32,
         capture_picker: bool,
     ) {
-        if !capture_picker && self.input_state.help_overlay.visible {
+        if !capture_picker && self.input_state.help_overlay.is_visible() {
             let bindings = crate::ui::HelpOverlayBindings::from_input_state(&self.input_state);
             let scroll_max = crate::ui::render_help_overlay(
                 ctx,
@@ -199,14 +199,14 @@ impl WaylandState {
                 width,
                 height,
                 self.frozen_enabled(),
-                self.input_state.help_overlay.page,
+                self.input_state.help_overlay.page(),
                 &bindings,
-                self.input_state.help_overlay.search.as_str(),
+                self.input_state.help_overlay.query(),
                 self.config.ui.help_overlay_context_filter,
                 self.input_state.boards.board_count() > 1,
                 self.config.capture.enabled,
-                self.input_state.help_overlay.scroll,
-                self.input_state.help_overlay.quick_mode,
+                self.input_state.help_overlay.scroll(),
+                self.input_state.help_overlay.is_quick_mode(),
             );
             self.input_state
                 .help_overlay
