+++ src/input/state/AGENTS.md
@@ -5,7 +5,7 @@
 
 ## Architecture
 - `actions/` owns action dispatch and key press/release behavior.
-- `core/` owns board state, history, selection, panels, properties, command palette, board picker, utilities, and session preflight. Panels with their own lifecycle own a state type (`core/font_picker/state.rs`, `core/command_palette/state.rs`); `core/base/ui_visibility.rs` groups the UI visibility preferences; `core/search.rs` holds the shared fuzzy scorer.
+- `core/` owns board state, history, selection, panels, properties, command palette, board picker, utilities, and session preflight. Panels with their own lifecycle own a state type (`core/font_picker/state.rs`, `core/command_palette/state.rs`, `core/help_overlay/state.rs`); `core/base/ui_visibility.rs` groups the UI visibility preferences; `core/search.rs` holds the shared fuzzy scorer.
 - `from_config.rs` is the only place configuration becomes an `InputState`; runtime code constructs through `InputState::from_config` and tests through `test_support::TestInputStateBuilder`.
 - `mouse/`, `interaction/`, and `highlight/` own pointer/mouse routing, interaction adapters, and highlight state.
 - `tests/` owns focused input state tests.
