+++ src/input/state/actions/help_overlay.rs
@@ -14,7 +14,7 @@ impl InputState {
             Key::Escape => {
                 // Escape clears search first, then closes overlay
                 if search_active {
-                    self.help_overlay.clear_search_text();
+                    self.help_overlay.clear_search();
                     self.dirty_tracker.mark_full();
                     self.needs_redraw = true;
                 } else {
@@ -27,7 +27,7 @@ impl InputState {
                 true
             }
             Key::Backspace if !self.help_overlay.search.is_empty() => {
-                self.help_overlay.pop_search_char();
+                self.help_overlay.backspace_search();
                 self.dirty_tracker.mark_full();
                 self.needs_redraw = true;
                 true
@@ -35,14 +35,16 @@ impl InputState {
             Key::Backspace => false,
             Key::Space => {
                 if search_active {
-                    self.help_overlay.push_search_char(' ');
+                    self.help_overlay.insert_search(" ");
                     self.dirty_tracker.mark_full();
                     self.needs_redraw = true;
                 }
                 true
             }
             Key::Char(ch) if !ch.is_control() => {
-                self.help_overlay.push_search_char(ch);
+                let mut encoded = [0; 4];
+                self.help_overlay
+                    .insert_search(ch.encode_utf8(&mut encoded));
                 self.dirty_tracker.mark_full();
                 self.needs_redraw = true;
                 true
@@ -54,8 +56,22 @@ impl InputState {
             {
                 true
             }
-            Key::Left | Key::PageUp if !search_active => self.help_overlay_prev_page(),
-            Key::Right | Key::PageDown if !search_active => self.help_overlay_next_page(),
+            Key::Left | Key::PageUp if !search_active => {
+                let changed = self.help_overlay.previous_page();
+                if changed {
+                    self.dirty_tracker.mark_full();
+                    self.needs_redraw = true;
+                }
+                changed
+            }
+            Key::Right | Key::PageDown if !search_active => {
+                let changed = self.help_overlay.next_page();
+                if changed {
+                    self.dirty_tracker.mark_full();
+                    self.needs_redraw = true;
+                }
+                changed
+            }
             Key::Home if !search_active => {
                 let changed = self.help_overlay.first_page();
                 if changed {
