+++ src/input/state/core/help_overlay/mod.rs
@@ -1,3 +1,6 @@
 pub(super) mod state;
 
 pub use state::HelpOverlayState;
+
+#[cfg(test)]
+mod tests;
